+++ autoscheduler/frontend/src/components/LandingPage/About/About.css.d.ts
@@ -1,15 +1,7 @@
 declare namespace AboutCssNamespace {
   export interface IAboutCss {
-    "close-button": string;
-    closeButton: string;
     "contributors-row": string;
     contributorsRow: string;
-    "dialog-content": string;
-    "dialog-link": string;
-    "dialog-title": string;
-    dialogContent: string;
-    dialogLink: string;
-    dialogTitle: string;
     "override-icon-button": string;
     overrideIconButton: string;
   }
